+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * `tableby()`: fixed a bug when some `Surv()` elements are NA. (#208)
 
+* `comparedf()`: added additional summary table to the `summary()` output.
+
 # arsenal v3.0.1 (local install only)
 
 * `tableby.control()`: fixed a bug with simplifying categorical and numeric output. (#199, #203)